Replace FourWinning game logic with Exact implementation and remove unused rendering code
